Rename example files
