+++ example/Receipt.tsx
